+++ src/gui/menu.ts
@@ -452,6 +452,7 @@ modmanager.gui.Menu = sc.ListInfoMenu.extend({
                 ig.input.pressed('special') /* space bar */ ||
                 /* R2 press */ ig.gamepad.isButtonPressed(sc.control._getSpecialButton())
         )
+
     },
     hideMenu(_afterSubmenu, nextSubmenu) {
         this.parent()
@@ -489,7 +490,7 @@ modmanager.gui.Menu = sc.ListInfoMenu.extend({
                 const main = sc.menu.guiReference
                 main.removeChildGui(this)
                 this.removeObservers()
-                main.submenus[modsMenuId]
+                delete main.submenus[modsMenuId]
             }, 1000)
         }
     },